Reenable viewBox coordinate transformation in the SVG parser
ViewBox trasnformation was commented out. This breaks all uses of non trivial
ViewBoxes - especially SVG files which coordinates do not match the document
settings.

Turns out the commented out code was actually wrong as it did scale the
coordinates first and then move the starting corner to the origin.
For this to work properly the starting corner needs to be move to the origin
first and coordinates must only scaled after that to ensure the starting corner
actually ends up on the origin.